+++ app/src/main/java/woowacourse/shopping/data/CartRepository.kt
@@ -2,19 +2,8 @@ package woowacourse.shopping.data
 
 import woowacourse.shopping.model.Product
 
-// TODO: Step2 - CartProductDao를 참조하도록 변경
-class CartRepository {
-
-    private val cartProducts: MutableList<Product> = mutableListOf()
-    fun addCartProduct(product: Product) {
-        cartProducts.add(product)
-    }
-
-    fun getAllCartProducts(): List<Product> {
-        return cartProducts.toList()
-    }
-
-    fun deleteCartProduct(id: Int) {
-        cartProducts.removeAt(id)
-    }
+interface CartRepository {
+    fun addCartProduct(product: Product)
+    fun getAllCartProducts(): List<Product>
+    fun deleteCartProduct(id: Int)
 }
